# internal/server/openapi.yaml
openapi: 3.0.3
info:
  title: VulnScanner REST API
  version: 1.0.0
  description: |
    VulnScanner 资产漏洞扫描与管理平台 REST API。
    - 鉴权：请求头 `X-API-Key`（admin 级自动化凭证）或用户登录返回的 `Authorization: Bearer <token>`（按角色鉴权），二者任选其一。
    - 角色：`admin` 全部操作；`operator` 日常运维写操作；`viewer` 只读（`/audit-logs*` 仅 admin 可见）。
    - 公开端点：`/health`、`/demo`、`/dl/*`、`/r/*`、`/api/v1/register`、`/api/v1/auth/login`、`/api/v1/auth/ldap/login`、`/openapi.yaml`。
    - 错误格式统一为 `{"error": "<message>"}`。
servers:
  - url: /
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /api/v1/auth/change-password:
    post:
      tags: [auth]
      summary: 修改本人密码
      x-roles: [admin, operator, viewer]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: 密码已更新
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: {type: string}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /api/v1/tenants:
    get:
      tags: [auth]
      summary: 租户列表（仅系统管理员）
      x-roles: [admin]
      responses:
        '200':
          description: 租户列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  tenants:
                    type: array
                    items:
                      $ref: '#/components/schemas/Tenant'
    post:
      tags: [auth]
      summary: 创建租户（仅系统管理员）
      x-roles: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TenantInput'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
        '400': {$ref: '#/components/responses/BadRequest'}
  /api/v1/users/{userId}/tenant:
    put:
      tags: [auth]
      summary: 调整用户所属租户（仅系统管理员）
      x-roles: [admin]
      parameters:
        - name: userId
          in: path
          required: true
          schema: {type: integer, format: int64}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenant_id]
              properties:
                tenant_id: {type: integer, format: int64}
      responses:
        '200':
          description: 调整成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id: {type: integer, format: int64}
                  tenant_id: {type: integer, format: int64}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
...
  /api/v1/audit-logs:
    get:
      tags: [audit]
      summary: 查询统一审计日志（仅 admin）
      x-roles: [admin]
      parameters:
        - {name: actor, in: query, required: false, schema: {type: string}, description: 操作人精确匹配}
        - {name: method, in: query, required: false, schema: {type: string}, description: HTTP 方法精确匹配}
        - {name: path, in: query, required: false, schema: {type: string}, description: 路径子串}
        - {name: since, in: query, required: false, schema: {type: string, format: date-time}, description: RFC3339 起始时间}
        - {name: until, in: query, required: false, schema: {type: string, format: date-time}, description: RFC3339 结束时间}
        - {name: limit, in: query, required: false, schema: {type: integer, default: 200, maximum: 1000}}
        - {name: offset, in: query, required: false, schema: {type: integer, default: 0}}
      responses:
        '200':
          description: 审计日志分页列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditLogList'
        '400': {$ref: '#/components/responses/BadRequest'}
...
  /api/v1/agents:
    get:
      tags: [agents]
      summary: Agent 列表
      x-roles: [admin, operator, viewer]
      responses:
        '200':
          description: Agent 数组
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Agent'
    post:
      tags: [agents]
      summary: 手动创建 Agent 并生成注册码
      x-roles: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [hostname, platform]
              properties:
                hostname: {type: string}
                platform: {type: string, enum: [linux-amd64, linux-arm64, windows-amd64, windows-arm64]}
      responses:
        '201':
          description: 创建成功，返回注册码与安装命令
          content:
            application/json:
              schema:
                type: object
                properties:
                  agent_id: {type: string}
                  code: {type: string}
                  install_cmd: {type: string}
        '400': {$ref: '#/components/responses/BadRequest'}
...
  /api/v1/patch-tasks:
    get:
      tags: [patch]
      summary: 全局补丁任务列表
      x-roles: [admin, operator, viewer]
      parameters:
        - {name: agent_id, in: query, required: false, schema: {type: string}}
        - {name: campaign_id, in: query, required: false, schema: {type: integer, format: int64}}
        - {name: status, in: query, required: false, schema: {type: string}}
        - {name: asset_name, in: query, required: false, schema: {type: string}}
        - {name: limit, in: query, required: false, schema: {type: integer, default: 100, maximum: 200}}
        - {name: offset, in: query, required: false, schema: {type: integer, default: 0}}
      responses:
        '200':
          description: 补丁任务列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks:
                    type: array
                    items:
                      $ref: '#/components/schemas/PatchTask'
                  total: {type: integer}
        '400': {$ref: '#/components/responses/BadRequest'}
...
  /api/v1/edr/findings:
    post:
      tags: [alerting]
      summary: 上报 EDR 恶意软件发现（幂等入库，HIGH/CRITICAL 自动生成告警）
      x-roles: [admin, operator]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EDRFindingInput'
      responses:
        '200':
          description: 已存在发现已刷新（同一去重键再次上报）
          content:
            application/json:
              schema:
                type: object
                properties:
                  finding:
                    $ref: '#/components/schemas/EDRFinding'
                  created: {type: boolean}
                  alerted: {type: boolean}
        '201':
          description: 新发现已入库
          content:
            application/json:
              schema:
                type: object
                properties:
                  finding:
                    $ref: '#/components/schemas/EDRFinding'
                  created: {type: boolean}
                  alerted: {type: boolean}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
    get:
      tags: [alerting]
      summary: 列出 EDR 恶意软件发现
      x-roles: [admin, operator, viewer]
      parameters:
        - {name: status, in: query, schema: {type: string, enum: [open, acknowledged, ignored, resolved]}}
        - {name: source, in: query, schema: {type: string}}
        - {name: agent_id, in: query, schema: {type: string}}
        - {name: severity, in: query, schema: {type: string, enum: [LOW, MEDIUM, HIGH, CRITICAL]}}
        - {name: q, in: query, schema: {type: string}}
        - {name: limit, in: query, schema: {type: integer}}
        - {name: offset, in: query, schema: {type: integer}}
      responses:
        '200':
          description: 发现列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  findings:
                    type: array
                    items:
                      $ref: '#/components/schemas/EDRFinding'
                  total: {type: integer}
...
  /api/v1/patch-campaigns:
    post:
      tags: [patch]
      summary: 按筛选批量生成补丁 Campaign（支持 dry-run）
      x-roles: [admin, operator]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GeneratePatchCampaignRequest'
      responses:
        '200':
          description: dry-run 演练结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  dry_run: {type: boolean}
                  matched: {type: integer}
                  counts:
                    $ref: '#/components/schemas/CampaignSummary'
                  errors:
                    type: array
                    items: {type: string}
        '201':
          description: Campaign 创建成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  campaign:
                    $ref: '#/components/schemas/PatchCampaign'
                  created: {type: integer}
                  counts:
                    $ref: '#/components/schemas/CampaignSummary'
                  tasks:
                    type: array
                    items:
                      $ref: '#/components/schemas/PatchTask'
                  errors:
                    type: array
                    items: {type: string}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
    get:
      tags: [patch]
      summary: Campaign 列表
      x-roles: [admin, operator, viewer]
      parameters:
        - {name: limit, in: query, required: false, schema: {type: integer, default: 100, maximum: 200}}
        - {name: offset, in: query, required: false, schema: {type: integer, default: 0}}
      responses:
        '200':
          description: Campaign 分页列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  campaigns:
                    type: array
                    items:
                      $ref: '#/components/schemas/PatchCampaign'
                  total: {type: integer}
...
  /api/v1/alert-rules:
    get:
      tags: [alerting]
      summary: 告警规则列表
      x-roles: [admin, operator, viewer]
      responses:
        '200':
          description: 告警规则
          content:
            application/json:
              schema:
                type: object
                properties:
                  rules:
                    type: array
                    items:
                      $ref: '#/components/schemas/AlertRule'
    post:
      tags: [alerting]
      summary: 创建告警规则
      x-roles: [admin, operator]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertRuleInput'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertRule'
        '400': {$ref: '#/components/responses/BadRequest'}
...
  /api/v1/alerts:
    get:
      tags: [alerting]
      summary: 告警列表
      x-roles: [admin, operator, viewer]
      parameters:
        - {name: status, in: query, required: false, schema: {type: string}}
        - {name: agent_id, in: query, required: false, schema: {type: string}}
        - {name: severity, in: query, required: false, schema: {type: string}}
        - {name: asset_filter, in: query, required: false, schema: {type: string}}
        - {name: limit, in: query, required: false, schema: {type: integer, default: 100}}
        - {name: offset, in: query, required: false, schema: {type: integer, default: 0}}
      responses:
        '200':
          description: 告警列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  alerts:
                    type: array
                    items:
                      $ref: '#/components/schemas/AlertDetail'
                  total: {type: integer}
...
  /api/v1/exceptions:
    get:
      tags: [risk]
      summary: 风险豁免列表
      x-roles: [admin, operator, viewer]
      responses:
        '200':
          description: 豁免列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  exceptions:
                    type: array
                    items:
                      $ref: '#/components/schemas/VulnerabilityException'
    post:
      tags: [risk]
      summary: 创建风险豁免
      x-roles: [admin, operator]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExceptionRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VulnerabilityException'
        '400': {$ref: '#/components/responses/BadRequest'}
...
  /api/v1/analyze:
    post:
      tags: [analysis]
      summary: 触发 LLM 分析
      x-roles: [admin, operator]
      responses:
        '202':
          description: 已受理（未配置 LLM 时 501）
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: {type: string}
                  analysis_id: {type: string}
        '501':
          description: LLM 未配置
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/v1/network/scan:
    post:
      tags: [scanning]
      summary: 创建网络扫描任务（Agent 领取执行）
      x-roles: [admin, operator]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NetworkScanRequest'
      responses:
        '201':
          description: 任务已创建
          content:
            application/json:
              schema:
                type: object
                properties:
                  task:
                    $ref: '#/components/schemas/NetworkScanTask'
        '400': {$ref: '#/components/responses/BadRequest'}
  /api/v1/remote/credentials:
    get:
      tags: [scanning]
      summary: 远程凭据列表（仅 admin，隐藏密码）
      x-roles: [admin]
      responses:
        '200':
          description: 凭据列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  credentials:
                    type: array
                    items:
                      $ref: '#/components/schemas/RemoteCredential'
    post:
      tags: [scanning]
      summary: 创建远程凭据（password/key）；key 不传私钥时服务生成密钥并返回公钥
      x-roles: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RemoteCredentialInput'
      responses:
        '201':
          description: 已创建，key 自生成时返回 public_key 仅一次
          content:
            application/json:
              schema:
                type: object
                properties:
                  credential:
                    $ref: '#/components/schemas/RemoteCredential'
                  public_key: {type: string}
        '400': {$ref: '#/components/responses/BadRequest'}
        '503':
          description: 远程扫描未启用或主密钥未配置
...
  /api/v1/remote/scan:
    post:
      tags: [scanning]
      summary: 创建远程 SSH 扫描任务，服务端直连目标采集
      x-roles: [admin, operator]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RemoteScanRequest'
      responses:
        '201':
          description: 任务已创建
          content:
            application/json:
              schema:
                type: object
                properties:
                  count: {type: integer}
                  tasks:
                    type: array
                    items:
                      $ref: '#/components/schemas/RemoteScanTask'
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
...
  /api/v1/cloud/accounts:
    get:
      tags: [cloud]
      summary: 云账号列表
      x-roles: [admin]
      responses:
        '200':
          description: 云账号
          content:
            application/json:
              schema:
                type: object
                properties:
                  accounts:
                    type: array
                    items:
                      $ref: '#/components/schemas/CloudAccount'
    post:
      tags: [cloud]
      summary: 创建云账号
      x-roles: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CloudAccountInput'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  account:
                    $ref: '#/components/schemas/CloudAccount'
        '400': {$ref: '#/components/responses/BadRequest'}
        '503':
          description: 云扫描未启用或主密钥未配置
...
  /api/v1/webdb/credentials:
    get:
      tags: [scanning]
      summary: Web/数据库扫描凭据列表
      x-roles: [admin]
      responses:
        '200':
          description: 凭据列表（不含密文）
          content:
            application/json:
              schema:
                type: object
                properties:
                  credentials:
                    type: array
                    items:
                      $ref: '#/components/schemas/WebDBCredential'
    post:
      tags: [scanning]
      summary: 创建 Web/数据库扫描凭据
      x-roles: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebDBCredentialInput'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  credential:
                    $ref: '#/components/schemas/WebDBCredential'
        '400': {$ref: '#/components/responses/BadRequest'}
        '503':
          description: Web/DB 扫描未启用或主密钥未配置
...
  /api/v1/webdb/scan:
    post:
      tags: [scanning]
      summary: 下发 Web 与数据库扫描任务
      x-roles: [admin, operator]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebDBScanRequest'
      responses:
        '201':
          description: 任务已创建
          content:
            application/json:
              schema:
                type: object
                properties:
                  count: {type: integer}
                  tasks:
                    type: array
                    items:
                      $ref: '#/components/schemas/WebDBTask'
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
        '503':
          description: Web/DB 扫描未启用或主密钥未配置
...
  /api/v1/container/scan:
    post:
      tags: [scanning]
      summary: 触发容器镜像扫描
      x-roles: [admin, operator]
      responses:
        '202':
          description: 扫描已触发
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: {type: string}
        '400': {$ref: '#/components/responses/BadRequest'}
...
  /api/v1/assets:
    get:
      tags: [assets]
      summary: 资产列表（筛选 + 分页）
      x-roles: [admin, operator, viewer]
      parameters:
        - {name: asset_type, in: query, required: false, schema: {type: string}}
        - {name: environment, in: query, required: false, schema: {type: string}}
        - {name: business_unit, in: query, required: false, schema: {type: string}}
        - {name: owner, in: query, required: false, schema: {type: string}}
        - {name: lifecycle, in: query, required: false, schema: {type: string}}
        - {name: agent_id, in: query, required: false, schema: {type: string}}
        - {name: tag, in: query, required: false, schema: {type: string}}
        - {name: q, in: query, required: false, schema: {type: string}}
        - {name: limit, in: query, required: false, schema: {type: integer, default: 100}}
        - {name: offset, in: query, required: false, schema: {type: integer, default: 0}}
      responses:
        '200':
          description: 资产分页列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  assets:
                    type: array
                    items:
                      $ref: '#/components/schemas/Asset'
                  total: {type: integer}
...
  /api/v1/assets/import:
    post:
      tags: [assets]
      summary: 导入外部 CMDB 资产（JSON 或 CSV）
      x-roles: [admin, operator]
      parameters:
        - {name: format, in: query, required: false, schema: {type: string, enum: [json, csv], default: json}}
        - {name: mode, in: query, required: false, schema: {type: string, enum: [merge, full], default: merge}}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportAssetsRequest'
          text/csv:
            schema: {type: string}
      responses:
        '200':
          description: 导入结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportResult'
        '400': {$ref: '#/components/responses/BadRequest'}
...
  /api/v1/assets/bulk-meta:
    post:
      tags: [assets]
      summary: 批量维护资产元数据
      x-roles: [admin, operator]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkMetaRequest'
      responses:
        '200':
          description: 批量更新结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated: {type: integer}
                  total_items: {type: integer}
        '400': {$ref: '#/components/responses/BadRequest'}
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: admin 级自动化凭证（等同 admin 角色）。
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 用户登录返回的 JWT，按用户角色鉴权。
  responses:
    BadRequest:
      description: 请求参数或请求体不合法
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: 缺少或无效的鉴权凭证
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Tenant:
      type: object
      properties:
        id: {type: integer, format: int64}
        name: {type: string}
        slug: {type: string}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    TenantInput:
      type: object
      required: [name, slug]
      properties:
        name: {type: string}
        slug: {type: string, pattern: '^[a-z0-9-]+$'}
...
    ChangePasswordRequest:
      type: object
      required: [current_password, new_password]
      properties:
        current_password: {type: string}
        new_password: {type: string, minLength: 8}
    Agent:
      type: object
      properties:
        id: {type: string}
        hostname: {type: string}
        os_type: {type: string}
        os_version: {type: string}
        arch: {type: string}
        agent_ver: {type: string}
        ip: {type: string}
        status: {type: string}
        tenant_id: {type: integer, format: int64}
        last_seen: {type: string, format: date-time}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
        eol_status: {type: string, enum: [eol, unsupported, supported, unknown]}
        eol_date: {type: string, format: date, nullable: true}
        eol_product: {type: string}
        eol_cycle: {type: string}
...
    AuditLogList:
      type: object
      properties:
        total: {type: integer}
        entries:
          type: array
          items:
            $ref: '#/components/schemas/AuditLog'
...
    PatchTask:
      type: object
      properties:
        id: {type: integer, format: int64}
        agent_id: {type: string}
        asset_name: {type: string}
        fix_type: {type: string}
        fix_value: {type: string}
        action: {type: string}
        cve_ids:
          type: array
          items: {type: string}
        command: {type: string}
        commands:
          type: array
          items:
            type: array
            items: {type: string}
        status: {type: string}
        approval_required: {type: boolean}
        window_start: {type: string, format: date-time, nullable: true}
        window_end: {type: string, format: date-time, nullable: true}
        cancel_requested: {type: boolean}
        runtime_verify_baseline:
          type: object
        runtime_verify_status: {type: string, enum: ['', pending, passed, failed, na]}
        runtime_verify_detail: {type: string}
        runtime_verify_at: {type: string, format: date-time, nullable: true}
        campaign_id: {type: integer, format: int64, nullable: true}
        result:
          type: object
        created_by: {type: string}
        approved_by: {type: string}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
...
    GeneratePatchCampaignRequest:
      type: object
      properties:
        name: {type: string}
        agent_ids:
          type: array
          items: {type: string}
        tags:
          type: array
          items: {type: string}
        environments:
          type: array
          items: {type: string}
        asset_names:
          type: array
          items: {type: string}
        cve_ids:
          type: array
          items: {type: string}
        min_severity: {type: string, enum: [LOW, MEDIUM, HIGH, CRITICAL]}
        min_cvss: {type: number, format: double}
        approval_required: {type: boolean}
        window_start: {type: string, format: date-time}
        window_end: {type: string, format: date-time}
        dry_run: {type: boolean}
    EDRFinding:
      type: object
      properties:
        id: {type: integer, format: int64}
        agent_id: {type: string}
        source: {type: string, enum: [edr_api, clamav]}
        finding_type: {type: string}
        name: {type: string}
        severity: {type: string, enum: [LOW, MEDIUM, HIGH, CRITICAL]}
        path: {type: string}
        hash: {type: string}
        detail: {type: string}
        status: {type: string, enum: [open, acknowledged, ignored, resolved]}
        occurrence_count: {type: integer}
        alert_id: {type: integer, format: int64, nullable: true}
        first_seen: {type: string, format: date-time}
        last_seen: {type: string, format: date-time}
    EDRFindingInput:
      type: object
      required: [name]
      properties:
        agent_id: {type: string}
        hostname: {type: string}
        source: {type: string, enum: [edr_api, clamav]}
        finding_type: {type: string, default: malware}
        name: {type: string}
        severity: {type: string, enum: [LOW, MEDIUM, HIGH, CRITICAL], default: MEDIUM}
        path: {type: string}
        hash: {type: string}
        detail: {type: string}
...